requirements.txt
alabaster==0.7.10
attrdict==2.0.0
awsrequests==0.0.12
Babel==2.5.3
boto3==1.7.11
botocore==1.10.11
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
cytoolz==0.9.0.1
docutils==0.14
emoji==0.5.0
eth-abi==1.0.0
eth-account==0.1.0a2
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0
eth-utils==1.0.3
Flask==0.12.2
gi==1.2
hexbytes==0.1.0
idna==2.6
imagesize==1.0.0
itsdangerous==0.24
jellyfish==0.6.1
Jinja2==2.10
jmespath==0.9.3
lru-dict==1.1.6
MarkupSafe==1.0
matplotlib==2.1.2
metapy==0.2.9
nltk==3.2.5
numpy==1.14.0
packaging==17.1
py-solc==2.1.0
pycryptodome==3.6.1
pyenchant==2.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-http-client==3.0.0
pytoml==0.1.14
pytz==2018.3
requests==2.18.4
rlp==0.6.0
s3transfer==0.1.13
scipy==1.0.0
semantic-version==2.6.0
sendgrid==5.3.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
textblob==0.15.1
toolz==0.9.0
urllib3==1.22
web3==4.1.0
websockets==4.0.1
Werkzeug==0.14.1
